pyproject.toml
[tool.poetry]
name = "django-breeze"
version = "0.1.0"
description = ""
authors = ["louxsdon <louisayivi.dev@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "django_breeze"}]

[tool.poetry.dependencies]
python = "^3.10"
django-vite = "^2.0.2"
inertia-django = "^0.5.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
